Add UI for timeline
Summary:
Add a GUI for the timeline of a Component based off MarkerTimeline.

It enables selecting points on the timeline, which will display their properties (arbitrary KVS). A button allows the user to select the current point of the timeline for the tree. This triggers unselection of the tree as per D24254002.

Solving re-selection isn't trivial, which is why the tree collapses on redraw too, and I believe we can live with this UX for this v1.

Reviewed By: astreet

Differential Revision: D23865369

fbshipit-source-id: 026562e71aa04b4e5e1e7cf1392d8ef140cdf43c