Add Android support for Timeline detail view
Summary:
This diff adds the Java classes that serialize to the data expected by the new timeline widget. See D23865369 for the JS counterpart.

Currently the Flipper plugin uses `toString` to get the wire values. I'm not feeling like using a json library or dealing with JSONObject's nonsense, so these are rolled manually for now.

Reviewed By: astreet

Differential Revision: D24254377

fbshipit-source-id: b2fc00400c40b47ac29c9b83c0b66621c3677974