useMemo -> useMemoize
Summary:
I love useMemo, and the memoized functions are in principle easy to testable since deps are known

But in practice they aren't as they close over their deps, preventing hoisting. So I now use a small `useMemoize` utility that encourages me to write them in a testable manner.

Applied this to the PluginList, and it makes it imho cleaner to read. Also, testing is now much easier as demonstrated in the next diff.

Reviewed By: nikoant

Differential Revision: D24445556

fbshipit-source-id: e156dbd4b8623eb1b90650e174b69a0ec868d2da