Support active sheet mechanism in Sandy
Summary:
This diffs adds support for the activeSheets mechanism in Sandy. It should be removed in the future (see T78696648) since open a dialog and keeping dialog state locally results in much more straight forward code, but supporting this for now makes sure that old flows are still supported.

With this change the plugin selection during a Flipper export for example wouldn't become visible

Reviewed By: cekkaewnumchai

Differential Revision: D24620074

fbshipit-source-id: f0558f5738e86a84a5cd0b9d574a3cfd0a3bf424