Support screencaps/recordings from multiple devices
Summary:
We've been using "booted" as a shortcut to select a device in `xcrun` but that will only refer to the *first* booted simulator. If you have multiple ones, that causes screenshots, videos and logs to always be pulled from the first.

Of course, this isn't documented anywhere, because Apple, but you can instead of `booted` pass in the "UDID" and refer to a specific device in all cases.

Reviewed By: nikoant

Differential Revision: D24627077

fbshipit-source-id: 2a52a8a17776746af4b3f6ca821dcce37bb0f7c2