Fix Empty Array And Object Not Showing up
Summary:
Reported in [the Workplace group](https://fb.workplace.com/groups/flippersupport/permalink/1002324710248187/), the empty array and object in network plugin wasn't shown and left blank. This happened because data component checked for expandability and decided empty object is not expandable. Thus, it uses different path to render components.

This diff fixes that by rendering empty array and object as `[]` and `{}`

Reviewed By: jknoxville

Differential Revision: D24726219

fbshipit-source-id: 9b22e1011c39c363b73f759477f44f08520734a7