Move public build infra entirely to GitHub
Summary:
As part of our release confidence effort, this replaces the flaky internal task that creates the tag, promotes it to a release and uploads the artifacts to it.

Instead, we do all this in one lengthy GitHub Action. There seems to be some duplication here but this is because of the unfortunate restriction GitHub Actions imposes to avoid recursive Actions. I.e. we cannot create a tag/release and have another, independent Action pick up from there to do the building and releasing.

The `passy/` references were necessary because existing Actions didn't quite what we wanted to:

- `tag-version-commit` expected the title of the commit to match the precise tag to create. However, we want to use a regex to *extract* from the "Flipper Release: vA.B.C". My fork allows specifying a grouping regex for that.
- `github-upload-release-artifacts-action` relied on being triggered by a tag. As described above, that's not possible with this setup. Instead, this takes an existing tag and attaches to it.

Reviewed By: nikoant

Differential Revision: D24627518

fbshipit-source-id: 8121df6aa7bd36bda28e0d3cb207a002cd127647