Use predefined icons for Facebook family apps
Summary:
Ideally, we want to fetch app Icons from the device, but for now, we can at least for Facebook owned apps we can use a hardcoded set of icons, which probably even looks better.

For ` Lite` apps, the colors of the icon and background are reversed, which makes both variations recognizable.

Reviewed By: jknoxville

Differential Revision: D24920161

fbshipit-source-id: 5220093fb3a443ae535faa4c346a2029de9cbbd5