Replace ITSR GK with a new one
Summary:
Michel discovered that the ITSR GK has been dialled down completely since the last bug.

Coming to re-enable it I discovered a bug that has been introduced since then, so have fixed that, but we can't just ramp up the original GK because the current version of flipper will crash if ITSR is used.

So instead, I'm leaving the current GK ramped down, and replacing it with a new one, which is safe to ramp up.

This will mean ITSR is only enabled on versions from the next release onwards.

Reviewed By: nikoant

Differential Revision: D24922822

fbshipit-source-id: 8a6990fd5130bd046c24b0da757492b37e4040f2