Move ITSR popup into Sandy LeftRail
Summary:
The title bar is going away, so this needs to be moved somewhere when sandy is enabled.

This moves it into the left rail when sandy is enabled, and converts the icon to use a sandy icon and popover.
The contents of the popover haven't yet been converted to use sandy components.

Reviewed By: mweststrate

Differential Revision: D25055282

fbshipit-source-id: 259ed8312c0d4079433cfa7ffb88385184ae16b2