Fix AdsManager plugin not loading
Summary:
FluxStores plugin of AdsManager didn't open on Sandy. The underlying reason is that in Sandy the list was based on the plugin id's as exposed in the `package.json`, while in the old left sidebar it was based on the `id` stored on the Plugin class. And apparently they can differ (will fix that in the next diff)

We can release an update of this plugin after it landed

Reviewed By: passy

Differential Revision: D25302357

fbshipit-source-id: 874aab263fad6a83f63edc4a8f7fa7ceceaef70f