Fix plugin meta data
Summary:
See previous diff, if the plugin `id` on the class definition mismatches with the `package.json`, the navigation in Sandy will use the wrong id (from the meta data) causing the plugin not to open.

Since the `id` exposed on the class is the one that is used to setup the connect on the client, updated the ids in the `package.json` which have afaik no further meaning.

Pinging nikoant to verify that assumption. I can imagine this might impact historical data / plugin distribution?

Reviewed By: passy, nikoant

Differential Revision: D25302542

fbshipit-source-id: 43fe1667604eb7a28f6f9b7d260bc2a1e6e82dd3