Add option to wipe local storage
Summary: This diff adds an option to wipe local storage to the settings dialog, to reset all preferences, table column widths, changelog and Nux state, etc. etc. in one go. This is primarily an self-fixing mechanism in case we mess up in the future, but useful during development / testing as well.

Reviewed By: passy

Differential Revision: D25421722

fbshipit-source-id: 3b69653b2142a7178ef83e01d35309f3a971e2cf