Delete duplicated extending layout plugin pages
Summary:
This content is currently duplicated in the fb internal section. So deleting it.
It's also in public here: https://www.internalfb.com/intern/staticdocs/flipper/docs/extending/layout-inspector

The only content in these files that isn't an exact dupe, is the iOS "Adding the plugin to a new app". But that's already covered in the plugin setup page: https://www.internalfb.com/intern/staticdocs/flipper/docs/setup/layout-plugin

Reviewed By: passy

Differential Revision: D25563321

fbshipit-source-id: 26d3b8c2c01c170cd440ae78bfaff20a13f90d37