Fix Screen Recording Not Working on Android Marshmallow
Summary: [Reported in the support group](https://fb.workplace.com/groups/flippersupport/permalink/1040131853134139/), screen recording can't be stopped on Android Marshmallow (SDK version 23). This is because the shell commands used in screen recording are different in SDK 23 compared to higher SDK. First, `pgrep` in SDK 23 doesn't have `-L` option. Second, `ls -l` of SDK 23 returns different pattern; we've relied on the 5th number to indicate the file size of a video, while it is on 4th number for SDK 23.

Reviewed By: passy

Differential Revision: D25641605

fbshipit-source-id: 110cb779ef4b2dc63a862bf5e4e5f3a304a0d8ec