Add github actions for dependency pod validations (#1834)
Summary:
This PR adds github actions to validate the pods, so that the changes which breaks the validation is not made. This PR also downgrades the version update of OpenSSL, as it doesn't validate the Folly podspec. I have [raised](https://github.com/krzyzanowskim/OpenSSL/issues/113) an issue on the repository which manages the openssl dependency.
## Changelog

- Added githubaction for dependent pod validation
- Downgraded the openssl version, as with upgraded version the folly pod didn't validate.

Pull Request resolved: https://github.com/facebook/flipper/pull/1834

Test Plan: CI

Reviewed By: passy

Differential Revision: D25939525

Pulled By: priteshrnandgaonkar

fbshipit-source-id: f70d043f1f6b2c79028212d8f56c6b8632092e33