Replace search view with drop down
Summary:
The support form currently has a search form to select a group, but unless the selection is cleared, it won't show you which groups is actually available, which makes it hard for people to select the right group if they don't know up front.

Since the scale of available groups doesn't justify needing a typeahead, converted it to an ordinary dropdown.

An added benefit is that this allows us to remove a large and complicated component we shouldn't be maintaining ourselves, but rather reuse from Ant.

Reviewed By: nikoant

Differential Revision: D26046131

fbshipit-source-id: f499e5848eec8b961b054104c8e3a01567e2801e