Split up the CK code so that internal refers to the latest and Open source refers to the sunset version and it won't be actively developed
Summary:
This diff keeps the opensource integration of CK for Flipper at the same place. It moves the code which is referred internally to a different folder under Facebook directory.

The reason we are doing this is because, CK won't be publishing any more opensource releases to Cocoapods. Thus we would like to keep the code for opensource compatible to the latest one. Internally we would refer to the CK code from Facebook directory.

Reviewed By: fabiomassimo

Differential Revision: D26047510

fbshipit-source-id: 302890bcd36d3e4c4888577d821433f58fcab5c7