Add gk checks in plugin
Summary:
This diff adds gk checks, VPN check and User login check to our plugin.

I added gk checks for Insta and FB apps. There are three kind of GK's. One with FB universe, one with Distillery universe and one with Insta WWW universe. We have API's to check gk with FB and Distillery universe. API to check with Insta WWW universe doesn't exists. There are some technical challenges to build this API. So right now for gks with Insta WWW universe, we just show an notification with info icon. For gks in other universe we aggressively check it when user closes the notif to verify if user assigned himself/herself to it.

In long term we are moving away from gks altogether, we need this UX improvement as it will help users to onboard easily.

See the following flow.

Reviewed By: mweststrate

Differential Revision: D26176996

fbshipit-source-id: 92a931610f9b244c14c6888bb12df936b62edd75