Update tutorial podfiles and GH actions (#1896)
Summary:
This PR updates the flipperkit version of the Tutorial and updates the GH action to test it without the arch flag.
This PR also fixes the script which updates the `flipperkit_version` tag in all the files. It was broken for Tutorial, as it used to replace the last released version of the pod with the current one in each file, but for some reason the flipper version in the Tutorial/Podfile was way older and it never got updated.

I have fixed the logic now to replace the current version in the file with the new one, rather than relying on only last version to be replaced.

## Changelog

- Updated the GH action of tutorial to remove the -arch flag
- Updated Tutorial/Podfile and Podfile.lock
- Updated scripts/update-pod-versions.sh

Pull Request resolved: https://github.com/facebook/flipper/pull/1896

Test Plan:
CI should be green
Tested that the version gets updated for each and every file, file tested were Flipper.podspec, FlipperKit.podspec, Sample/Podfile SampleSwift/Podfile Tutorial/Podfile

{F369967362}

{F369967395}

Reviewed By: nikoant

Differential Revision: D26275657

Pulled By: priteshrnandgaonkar

fbshipit-source-id: 0f21a2a33c740938d2b29624e95974e80a2e38ba