Set up release automation (#1915)
Summary:
Depends on https://github.com/facebook/flipper/issues/1914.
This sets up the workflows for automatically building, publishing and "closing" the repository on Maven Central, which is a step you usually need to do manually in the web UI.

Still need to transfer all the secrets over to this repo for it to work.

Pull Request resolved: https://github.com/facebook/flipper/pull/1915

Test Plan:
Ran all of this in my this in my fork and released some `0.0.x` versions.

- Successful workflow: https://github.com/passy/flipper-1/actions/runs/552170713
- It's actually all on Maven: https://repo.maven.apache.org/maven2/com/facebook/flipper/

Reviewed By: jknoxville

Differential Revision: D26367469

Pulled By: passy

fbshipit-source-id: d65f64ee8eec1dc71af64b7e0f87473b9c956fec