Dedupe errors when sending
Summary: Keeps a small ring buffer of errors we're sending and checks if they're tracked. I'm sure there's a better data structure for this out there were we could have O(1) membership checks, but since we're not keeping that many errors to begin with and we're saving a lot of CPU that would go into sending the errors to Scribe, this should be a net positive.

Reviewed By: mweststrate

Differential Revision: D26488825

fbshipit-source-id: fdcf58e88d43864852c13e4741a7ab620d596094