Add more gradle sub projects (#1997)
Summary:
I don't fully understand how dependabot works with Android but it seems to only look at the literal fbjni version number based on the logs and we're actually outdated on a bunch of dependencies which Gradle doesn't handle very well.

Pull Request resolved: https://github.com/facebook/flipper/pull/1997

Test Plan: Tested on my fork and I've got a ton of PRs because of this. Brace yourselves.

Reviewed By: mweststrate

Differential Revision: D26784112

Pulled By: passy

fbshipit-source-id: 98ebd0b899a0603868d4b6bc0866e557c87fecd5